Moved setFilter ir clearFilter to FilterIcon
